requirements.txt
# Core AI Framework
google-adk>=0.1.0
google-generativeai>=0.3.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Environment Management
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Kubernetes
kubernetes>=28.1.0

# Monitoring
prometheus-client>=0.19.0
